+++ NEWS.md
@@ -1,5 +1,9 @@
 # ospsuite 11.0 (development version)
 
+## New features
+
+* Adds `simulationResultsToDataFrame()` function to convert `SimulationResults` objects into a dataframe.
+
 ## Breaking changes
 
 * The `pkAnalysesAsDataFrame()` function changes column data types for `QuantityPath`, `Parameter`, and `Unit`  from `factor` to `character` (https://github.com/Open-Systems-Pharmacology/OSPSuite-R/issues/673).